# contracts/doorstep/openapi.yaml
openapi: 3.0.3
info:
  title: Doorstep Home Services API
  version: 1.0.0
  description: >-
    Authoritative REST contract for Doorstep (Urban Company style home
    services): the customer surface (Momentum Android app and postbook-ui
    /doorstep), the professional surface (/pro, the doorstep_pro Android app)
    and the admin-internal family (/internal/admin, admin-service tokens only).


    Served by doorstep-service (port 8122, schema `doorstep`) behind the API
    gateway prefix /v1/doorstep (dormant gate DOORSTEP_PUBLIC_ENABLED, pilot
    allowlist DOORSTEP_PILOT_USER_IDS). Pilot city Hyderabad (`HYD`).


    Conventions (binding for every lane):

    - Every response is the shared envelope: success `{"data": <payload>,
      "meta": {"request_id": "..."}}`, failure `{"error": {"code", "message",
      "details"?}, "meta": {...}}`. Schemas below describe `data` unless they
      are named `*Envelope` or `ErrorEnvelope`.

    - Money is integer paise in fields ending `_paise`. Customer prices are
      GST-inclusive; the split is shown as `taxable_paise` + `tax_paise`.

    - Error codes are stable UPPER_SNAKE strings (see `x-doorstep-error-codes`).
      Clients branch on `error.code`, never on `message`.

    - Timestamps are RFC 3339 UTC. Dates are `YYYY-MM-DD` in the city's time
      zone (Asia/Kolkata). Times of day are `HH:MM`.

    - Optional values are emitted as explicit `null`, never omitted, on every
...

...
      gateway refuses /internal/ from the edge.


...


...
servers:
...
security:
  - gatewayIdentity: []
...
paths:
...
  /v1/doorstep/addresses:
    get:
      summary: My saved addresses
      operationId: listAddresses
      tags: [Customer]
      x-lane: A3
      description: Default first, then newest. Street lines are sealed at rest (shared/pii) and opened for the owner only. Golden bytes addresses_get_200.json.
      responses:
        '200':
          description: Addresses
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items: {$ref: '#/components/schemas/Address'}
    post:
      summary: Save an address (point must be serviceable)
      operationId: createAddress
      tags: [Customer]
      x-lane: A3
      description: >-
        Serviceability on save: the point must be in an active zone (city and zone are derived
        from it). The first address is the default. line1, line2 and landmark are sealed. Golden
        bytes address_post_201.json, address_post_422_outside_area.json. 503
        DOORSTEP_PII_UNAVAILABLE when sealing keys are absent (development only).
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/AddressInput'}
      responses:
        '201':
          description: Saved
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Address'}
        '422':
          description: DOORSTEP_OUTSIDE_SERVICE_AREA
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorEnvelope'}
  /v1/doorstep/addresses/{id}:
    patch:
      summary: Edit an address
      operationId: updateAddress
      tags: [Customer]
      x-lane: A3
      description: Partial update. An absent field keeps its value (lat and lng change together); the point is re-checked for serviceability. Bookings keep their own snapshot.
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/AddressInput'}
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Address'}
        '404':
          $ref: '#/components/responses/Error'
    delete:
      summary: Remove an address (soft delete; bookings keep their snapshot)
      operationId: deleteAddress
      tags: [Customer]
      x-lane: A3
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '204':
          description: Removed
...
  /v1/doorstep/bookings:
    post:
      summary: Book the picked professional for a slot or as soon as possible (hold, then pay)
      operationId: createBooking
      tags: [Booking]
      x-lane: A3
      description: >-
        B1: the hold goes on the quote's professional (the customer's pick)
        and nobody else; on a lost race DOORSTEP_SLOT_TAKEN and the client
        lists again. Send exactly one of slot_start (the grid) or asap=true
        (an on-duty professional in range who opted in to same-day jobs: the
        block runs from now, the job starts at now + ETA, the hold lasts 5
        minutes, the offer 3 minutes; booking_post_201_asap.json). Every
        line's price must still be the professional's approved, live price
        (else DOORSTEP_PRICE_UNAVAILABLE). Inserts a `hold`
        calendar block (EXCLUDE USING gist makes a double hold impossible),
        creates the booking in pending_payment and the payments-service
        intent (application doorstep, reference doorstep_booking). Idempotent
        on the Idempotency-Key header per customer. A replay with the same body answers the
        same booking (and opens the intent if it was never opened); the same key with another
        quote, address or slot (or asap) is 400. The professional is reserved (reserved_pro_id).
        The hold is [slot_start, slot_end + the zone travel buffer) (ASAP: from the booking
        time). Outstanding extras refuse with 409. Golden bytes
        booking_post_201.json, booking_post_400_idempotency_key.json,
        booking_post_409_slot_taken.json, booking_post_409_outstanding.json,
        booking_post_410_quote_expired.json, booking_post_422_slot_unavailable.json
        (details.reason inside_lead_time, beyond_horizon, not_on_grid, no_professional,
        professional_unavailable or not_available_now).
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/BookingCreateRequest'}
      responses:
        '201':
          description: Booking created; pay before hold_expires_at
          content:
            application/json:
              schema: {$ref: '#/components/schemas/BookingCreated'}
        '409':
          description: DOORSTEP_SLOT_TAKEN, DOORSTEP_OUTSTANDING_DUE
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorEnvelope'}
        '410':
          description: DOORSTEP_QUOTE_EXPIRED
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorEnvelope'}
        '422':
          description: DOORSTEP_SLOT_UNAVAILABLE, DOORSTEP_OUTSIDE_SERVICE_AREA, DOORSTEP_PRICE_UNAVAILABLE
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorEnvelope'}
        '503':
          description: DOORSTEP_PAYMENTS_UNAVAILABLE
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorEnvelope'}
    get:
      summary: My bookings
      operationId: listBookings
      tags: [Booking]
      x-lane: A3
      description: upcoming = active bookings soonest first; past = finished, latest first; all = latest slot first. Golden bytes bookings_get_200.json.
      parameters:
        - name: status
          in: query
          description: upcoming | past | all
          schema: {type: string, enum: [upcoming, past, all], default: all}
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Page of bookings
          content:
            application/json:
              schema:
                type: object
                required: [items, next_cursor]
                properties:
                  items:
                    type: array
                    items: {$ref: '#/components/schemas/BookingSummary'}
                  next_cursor: {type: string, nullable: true}
  /v1/doorstep/bookings/{id}:
    get:
      summary: One of my bookings
      operationId: getBooking
      tags: [Booking]
      x-lane: A3
      description: Golden bytes booking_get_200.json, booking_get_200_pending_payment.json, booking_get_404.json (also for another customer's booking).
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Booking
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Booking'}
        '404':
          $ref: '#/components/responses/Error'
...
  /v1/doorstep/bookings/{id}/cancel:
    post:
      summary: Cancel my booking
      operationId: cancelBooking
      tags: [Booking]
      x-lane: A3
      description: Releases the calendar, cancels open offers, writes the refund (paid - fee, key doorstep:refund:{id}:customer_cancel) with the cancellation and submits it. Golden bytes booking_cancel_post_200.json.
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/CancelRequest'}
      responses:
        '200':
          description: Cancelled; refund (total - fee) requested
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Booking'}
        '409':
          description: DOORSTEP_CANCEL_NOT_ALLOWED, DOORSTEP_INVALID_TRANSITION
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorEnvelope'}
  /v1/doorstep/bookings/{id}/reschedule:
    post:
      summary: Move my booking to another slot (free once, up to 3 h before)
      operationId: rescheduleBooking
      tags: [Booking]
      x-lane: A3
      description: >-
        From confirmed or assigned, once, at least 3 h before the slot, with the same professional
        only (B1: the customer picked them and their price; a slot they cannot take is 422
        DOORSTEP_SLOT_UNAVAILABLE, nobody else is held). New block then the old one released, in
        one transaction. Golden bytes booking_reschedule_post_200.json,
        booking_reschedule_post_409.json.
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [slot_start]
              properties:
                slot_start: {type: string, format: date-time}
      responses:
        '200':
          description: Rescheduled (may re-run dispatch)
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Booking'}
        '409':
          description: DOORSTEP_RESCHEDULE_NOT_ALLOWED, DOORSTEP_SLOT_TAKEN
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorEnvelope'}
...
  /v1/doorstep/internal/admin/bookings:
    get:
      summary: Bookings (filters)
      operationId: adminListBookings
      tags: [Admin]
      security: [{adminServiceToken: []}]
      x-permission: doorstep:bookings.read
      x-lane: A3
      description: Latest slot first, 50 per page; rows carry no address. Golden bytes admin_bookings_list_200.json.
      parameters:
        - name: status
          in: query
          schema: {$ref: '#/components/schemas/BookingStatus'}
        - $ref: '#/components/parameters/CityFilter'
        - name: date
          in: query
          schema: {type: string, format: date}
        - $ref: '#/components/parameters/Cursor'
      responses:
        '200':
          description: Page
          content:
            application/json:
              schema:
                type: object
                required: [items, next_cursor]
                properties:
                  items:
                    type: array
                    items: {$ref: '#/components/schemas/BookingSummary'}
                  next_cursor: {type: string, nullable: true}
  /v1/doorstep/internal/admin/bookings/{id}:
    get:
      summary: Booking with history, assignments, payments, refunds, extras, photos
      operationId: adminGetBooking
      tags: [Admin]
      security: [{adminServiceToken: []}]
      x-permission: doorstep:bookings.read
      x-lane: A3
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Detail
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AdminBookingDetail'}
  /v1/doorstep/internal/admin/bookings/{id}/cancel:
    post:
      summary: Ops cancel (full refund unless fee_paise given)
      operationId: adminCancelBooking
      tags: [Admin]
      security: [{adminServiceToken: []}]
      x-permission: doorstep:bookings.cancel
      x-lane: A3
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/AdminCancelInput'}
      responses:
        '200':
          description: Booking (admin view, OTPs null). Golden bytes admin_booking_cancel_200.json. The refund (refundable - fee) is written with the cancellation, keyed doorstep:refund:{id}:admin_cancel.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Booking'}
...
  /v1/doorstep/internal/admin/bookings/{id}/refund:
    post:
      summary: Issue a refund (step-up in admin-service)
      operationId: adminRefundBooking
      tags: [Admin]
      security: [{adminServiceToken: []}]
      x-permission: doorstep:refunds.issue
      x-lane: A3
      description: >-
        Refunds the booking payment: amount_paise must not exceed captured minus every refund
        that has not failed (DOORSTEP_REFUND_EXCEEDS_PAID). Idempotent on the Idempotency-Key
        admin-service forwards (its two-person approval replays the same key): the refund key is
        doorstep:refund:{id}:admin_<hash of the key>, and a replay answers the same row. Written
        with its audit row in one transaction, then submitted to payments-service; settled only by
        payment.refunded. payment=extras answers 422 until the visit lane pays extras.
        Golden bytes admin_booking_refund_201.json, admin_booking_refund_422_exceeds.json.
      parameters:
        - $ref: '#/components/parameters/IdPath'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [amount_paise, reason]
              properties:
                amount_paise: {type: integer, format: int64, minimum: 1}
                reason: {type: string}
                payment: {type: string, enum: [booking, extras], default: booking}
      responses:
        '201':
          description: Refund requested
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Refund'}
        '400':
          description: DOORSTEP_INVALID_REQUEST (no Idempotency-Key, amount, reason)
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorEnvelope'}
        '422':
          description: DOORSTEP_REFUND_EXCEEDS_PAID
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorEnvelope'}

  # ---------- B1: professionals' price review ----------
...
  /v1/doorstep/internal/admin/stats:
    get:
      summary: Dashboard counts
      operationId: adminStats
      tags: [Admin]
      security: [{adminServiceToken: []}]
      x-permission: doorstep:stats.read
      x-lane: A3
      responses:
        '200':
          description: Counts (money in paise)
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AdminStats'}
...
components:
  securitySchemes:
    gatewayIdentity:
      type: apiKey
      in: header
      name: X-Verified-User-Id
      description: Stamped by the gateway from the verified JWT (with X-Internal-Service-Key). Never sent by clients directly.
    adminServiceToken:
      type: apiKey
      in: header
      name: X-Service-Authorization
      description: 'Bearer <ed25519 service token>: iss admin-service, aud doorstep, scope [one doorstep:* permission], act <admin user id>, exp <= 60 s.'
  parameters:
    IdPath:
      name: id
      in: path
      required: true
      schema: {type: string, format: uuid}
...
    CityFilter:
      name: city
      in: query
      required: false
      schema: {type: string, example: HYD}
    Cursor:
      name: cursor
      in: query
      schema: {type: string}
    Limit:
      name: limit
      in: query
      schema: {type: integer, minimum: 1, maximum: 100, default: 50}
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: {type: string, maxLength: 128}
  responses:
    Error:
      description: Error envelope (see x-doorstep-error-codes)
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ErrorEnvelope'}
...
  schemas:
...
    Meta:
      type: object
      properties:
        request_id: {type: string}
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: {type: string, example: DOORSTEP_ADDON_INVALID}
            message: {type: string}
            details:
              type: object
              additionalProperties: true
        meta: {$ref: '#/components/schemas/Meta'}

    # ---------- enums ----------
...
    BookingStatus:
      type: string
      enum: [pending_payment, confirmed, assigned, en_route, arrived, in_progress, awaiting_extras_payment, completed, cancelled, expired, customer_no_show, pro_no_show, pro_unavailable]
...
    Address:
      type: object
      required: [id, label, line1, line2, landmark, locality, city_code, pincode, lat, lng, zone_id, is_default, created_at]
      properties:
        id: {type: string, format: uuid}
        label: {type: string, example: Home}
        line1: {type: string}
        line2: {type: string, nullable: true}
        landmark: {type: string, nullable: true}
        locality: {type: string, description: The only part shown to a professional before acceptance}
        city_code: {type: string}
        pincode: {type: string}
        lat: {type: number}
        lng: {type: number}
        zone_id: {type: string, format: uuid}
        is_default: {type: boolean}
        created_at: {type: string, format: date-time}
    AddressInput:
      type: object
      required: [label, line1, locality, pincode, lat, lng]
      properties:
        label: {type: string}
        line1: {type: string}
        line2: {type: string}
        landmark: {type: string}
        locality: {type: string}
        pincode: {type: string}
        lat: {type: number}
        lng: {type: number}
        is_default: {type: boolean}
...
    BookingCreateRequest:
      type: object
      description: Exactly one of slot_start or asap=true (B1).
      required: [quote_id, address_id]
      properties:
        quote_id: {type: string, format: uuid}
        address_id: {type: string, format: uuid}
        slot_start: {type: string, format: date-time}
        asap: {type: boolean, description: "B1: same day, as soon as possible (the quote's professional on duty in range, opted in to same-day jobs)"}
        require_female_pro: {type: boolean, default: false}
        notes: {type: string, maxLength: 500}
    BookingCreated:
      type: object
      required: [booking, payment_intent]
      properties:
        booking: {$ref: '#/components/schemas/Booking'}
        payment_intent: {$ref: '#/components/schemas/PaymentIntent'}
    BookingSummary:
      type: object
      required: [id, status, service_name, category_slug, slot_start, slot_end, total_paise, created_at]
      properties:
        id: {type: string, format: uuid}
        status: {$ref: '#/components/schemas/BookingStatus'}
        service_name: {type: string}
        category_slug: {type: string}
        slot_start: {type: string, format: date-time}
        slot_end: {type: string, format: date-time}
        total_paise: {type: integer, format: int64}
        created_at: {type: string, format: date-time}
...
    Booking:
      type: object
      description: >-
        Golden bytes booking_get_200.json (confirmed), booking_get_200_pending_payment.json,
        booking_cancel_post_200.json, booking_reschedule_post_200.json. end_otp, photos and
        status_history closed the A3 contract gaps: end_otp stays null until the visit lane
        (A5) sets it, photos stays [] until A5 uploads before/after photos (customers see
        only those two phases), status_history is the timeline oldest first.
      required: [id, status, service_id, service_name, category_slug, city_code, zone_id, slot_start, slot_end, duration_minutes, require_female_pro, items, total_paise, taxable_paise, tax_paise, paid_paise, refunded_paise, cancellation_fee_paise, extras_total_paise, outstanding_paise, hold_expires_at, address, professional, parent_booking_id, start_otp, end_otp, photos, status_history, can_cancel, can_reschedule, asap, choice_deadline, unavailable_cause, pending_change, created_at, updated_at]
      properties:
        id: {type: string, format: uuid}
        status: {$ref: '#/components/schemas/BookingStatus'}
        service_id: {type: string, format: uuid}
        service_name: {type: string}
        category_slug: {type: string}
        city_code: {type: string}
        zone_id: {type: string, format: uuid}
        slot_start: {type: string, format: date-time}
        slot_end: {type: string, format: date-time}
        duration_minutes: {type: integer}
        require_female_pro: {type: boolean}
        items:
          type: array
          items: {$ref: '#/components/schemas/QuoteLine'}
        total_paise: {type: integer, format: int64}
        taxable_paise: {type: integer, format: int64}
        tax_paise: {type: integer, format: int64}
        paid_paise: {type: integer, format: int64}
        refunded_paise: {type: integer, format: int64}
        cancellation_fee_paise: {type: integer, format: int64}
        extras_total_paise: {type: integer, format: int64}
        outstanding_paise: {type: integer, format: int64}
        hold_expires_at: {type: string, format: date-time, nullable: true}
        address: {$ref: '#/components/schemas/Address'}
        professional:
          type: object
          nullable: true
          description: "B1: the professional the customer picked (from booking on), else the one who accepted; null while nobody holds the job (pro_unavailable). First name, photo, rating; a phone never."
          required: [first_name, photo_media_id, rating_avg, jobs_completed]
          properties:
            first_name: {type: string}
            photo_media_id: {type: string, nullable: true}
            rating_avg: {type: number, nullable: true}
            jobs_completed: {type: integer}
        parent_booking_id: {type: string, format: uuid, nullable: true}
        start_otp: {type: string, nullable: true, description: Shown to the customer from assigned until in_progress (A5 sets it; null before)}
        end_otp: {type: string, nullable: true, description: Shown to the customer while in_progress once the after photos are in (A5 sets it; null before). Never in an admin response.}
        photos:
          type: array
          description: Before/after visit photos (A5); [] until then
          items: {$ref: '#/components/schemas/Photo'}
        status_history:
          type: array
          items: {$ref: '#/components/schemas/StatusStep'}
        can_cancel: {type: boolean}
        can_reschedule: {type: boolean, description: "confirmed or assigned, never rescheduled before, and at least 3 h before the slot (same professional only)"}
        asap: {type: boolean, description: "B1: a same-day, as-soon-as-possible booking"}
        choice_deadline: {type: string, format: date-time, nullable: true, description: "B1: while pro_unavailable, pick another professional (or cancel) by then, else it is cancelled with a full refund"}
        unavailable_cause: {type: string, nullable: true, enum: [declined, offer_expired, pro_cancel, not_on_duty, pro_no_show, ops_redispatch, no_professional, null], description: "B1: why the professional is gone (pro_unavailable only)"}
        pending_change:
          allOf: [{$ref: '#/components/schemas/ProChange'}]
          nullable: true
          description: "B1: a change of professional waiting for its difference to be paid (with the payment intent), else null"
        created_at: {type: string, format: date-time}
        updated_at: {type: string, format: date-time}
    CancelRequest:
      type: object
      required: [reason]
      properties:
        reason: {type: string, maxLength: 500}
...
    Refund:
      type: object
      required: [id, payment_id, cause, amount_paise, status, created_at]
      properties:
        id: {type: string, format: uuid}
        payment_id: {type: string, format: uuid}
        cause: {type: string, example: customer_cancel}
        amount_paise: {type: integer, format: int64}
        status: {type: string, enum: [requested, pending, succeeded, failed]}
        created_at: {type: string, format: date-time}
...
    Photo:
      type: object
      required: [id, booking_id, phase, media_id, created_at]
      properties:
        id: {type: string, format: uuid}
        booking_id: {type: string, format: uuid}
        phase: {type: string, enum: [before, after, kit_seal, extra_evidence]}
        media_id: {type: string, format: uuid}
        created_at: {type: string, format: date-time}
...
    AdminCancelInput:
      type: object
      required: [reason]
      properties:
        reason: {type: string, maxLength: 1000}
        fee_paise:
          type: integer
          format: int64
          minimum: 0
          description: Omitted = full refund of what is refundable; a fee is capped at it
...
    AdminStats:
      type: object
      required: [bookings_today, bookings_in_progress, unassigned_within_2h, bookings_needing_attention, pros_approved, pros_pending_verification, documents_pending, incidents_open, gmv_today_paise, refunds_today_paise, outstanding_paise]
      description: Golden bytes admin_stats_200.json. "Today" is the IST day.
      properties:
        bookings_today: {type: integer}
        bookings_in_progress: {type: integer}
        unassigned_within_2h: {type: integer}
        bookings_needing_attention: {type: integer}
        pros_approved: {type: integer}
        pros_pending_verification: {type: integer}
        documents_pending: {type: integer}
        incidents_open: {type: integer}
        gmv_today_paise: {type: integer, format: int64}
        refunds_today_paise: {type: integer, format: int64}
        outstanding_paise: {type: integer, format: int64}
...